Added func clear_font_outline(), but result could be better
Former-commit-id: 6100c5ab4c572301945ec61c59f3b26f3820dd62

+++ src_bonus/controller.c
@@ -35,8 +35,8 @@ void	change_textures(t_game *game)
 	int			counter;
 	int			frames_to_move;
 
-	frames_to_move = (float)(get_time() - game->time.startup) / clocks_per_frame
-		- (float)(game->time.last - game->time.startup) / clocks_per_frame;
+	frames_to_move = (int)((float)(get_time() - game->time.startup) / clocks_per_frame)
+		- (int)((float)(game->time.last - game->time.startup) / clocks_per_frame);
 	if (frames_to_move)
 	{
 		index = -1;
